requirements.txt
requests
python-qbittorrent
psutil
SpeechRecognition
pyaudio
pony